DIGISOS-159-haandtere-serverfeil * Flytt nav-soknad/redux/fakta* inn i sin egen mappe * Flytt typedefinisjoner ut av faktaReducer og inn i faktaTypes
