+++ web/src/frontend/src/digisos/redux/reducers.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from "redux";
 
 import { SoknadAppState } from "../../nav-soknad/redux/reduxTypes";
-import FaktumReducer from "../../nav-soknad/redux/faktaReducer";
+import FaktumReducer from "../../nav-soknad/redux/fakta/faktaReducer";
 import ValideringReducer from "../../nav-soknad/redux/valideringReducer";
 import OppsummeringReducer from "../../nav-soknad/redux/oppsummeringReducer";
 import SoknadReducer from "../../nav-soknad/redux/soknadReducer";
