+++ web/src/frontend/src/digisos/skjema/ekstrainformasjon/Opplysning.tsx
@@ -6,9 +6,9 @@ import { Column, Container, Row } from "nav-frontend-grid";
 import BelopFaktum from "../../../nav-soknad/faktum/typedInput/BelopFaktum";
 import { DispatchProps, SoknadAppState } from "../../../nav-soknad/redux/reduxTypes";
 import { Faktum } from "../../../nav-soknad/types/navSoknadTypes";
-import { FaktumComponentProps } from "../../../nav-soknad/redux/faktaReducer";
+import { FaktumComponentProps } from "../../../nav-soknad/redux/fakta/faktaTypes";
 import { finnFakta, finnFaktum } from "../../../nav-soknad/utils/faktumUtils";
-import { opprettFaktum, slettFaktum } from "../../../nav-soknad/redux/faktaActions";
+import { opprettFaktum, slettFaktum } from "../../../nav-soknad/redux/fakta/faktaActions";
 import { injectIntl, InjectedIntlProps } from "react-intl";
 import Lenkeknapp from "../../../nav-soknad/components/lenkeknapp/Lenkeknapp";
 import InputFaktum from "../../../nav-soknad/faktum/InputFaktum";
