+++ web/src/frontend/src/digisos/skjema/familie/HarBarn.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import SporsmalFaktum from "../../../nav-soknad/faktum/SporsmalFaktum";
 import FaktumRadio from "../../../nav-soknad/faktum/RadioFaktum";
 import Underskjema from "../../../nav-soknad/components/underskjema";
-import { FaktumComponentProps } from "../../../nav-soknad/redux/faktaReducer";
+import { FaktumComponentProps } from "../../../nav-soknad/redux/fakta/faktaTypes";
 import { getFaktumVerdi, radioCheckKeys } from "../../../nav-soknad/utils";
 
 import Barneinfo from "./Barneinfo";
