+++ web/src/frontend/src/digisos/skjema/familie/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { connect } from "react-redux";
-import { FaktumComponentProps } from "../../../nav-soknad/redux/faktaReducer";
+import { FaktumComponentProps } from "../../../nav-soknad/redux/fakta/faktaTypes";
 import { State } from "../../redux/reducers";
 import DigisosSkjemaSteg, { DigisosSteg } from "../DigisosSkjemaSteg";
