+++ web/src/frontend/src/digisos/skjema/utgifterGjeld/index.tsx
@@ -15,7 +15,7 @@ import {
 	radioCheckKeys
 } from "../../../nav-soknad/utils";
 import { DispatchProps } from "../../../nav-soknad/redux/reduxTypes";
-import { FaktumComponentProps } from "../../../nav-soknad/redux/faktaReducer";
+import { FaktumComponentProps } from "../../../nav-soknad/redux/fakta/faktaTypes";
 import { getMaksLengdeFunc } from "../../../nav-soknad/validering/valideringer";
 
 import DigisosSkjemaSteg, { DigisosSteg } from "../DigisosSkjemaSteg";
