+++ web/src/frontend/src/digisos/start/Bosted.tsx
@@ -5,7 +5,7 @@ import { injectIntl, InjectedIntlProps, FormattedMessage } from "react-intl";
 import Knapp from "nav-frontend-knapper";
 import { Select } from "nav-frontend-skjema";
 
-import { FaktumComponentProps } from "../../nav-soknad/redux/faktaReducer";
+import { FaktumComponentProps } from "../../nav-soknad/redux/fakta/faktaTypes";
 import Arrow from "../../nav-soknad/components/svg/Arrow";
 import { DispatchProps } from "../../nav-soknad/redux/reduxTypes";
 import { REST_STATUS } from "../../nav-soknad/types";
