+++ web/src/frontend/src/nav-soknad/redux/fakta/faktaReducer.ts
@@ -1,11 +1,8 @@
-import { REST_STATUS, Faktum } from "../types";
-import { FaktumActionTypeKeys, FaktaActionTypeKeys } from "./faktaActionTypes";
-import {
-	updateFaktumMedLagretVerdi,
-	updateFaktumStateVerdi,
-	updateFaktumLagretVerdi
-} from "./faktaUtils";
-import { Reducer } from "./reduxTypes";
+import { Faktum, REST_STATUS } from "../../types";
+import { FaktaActionTypeKeys, FaktumActionTypeKeys } from "./faktaActionTypes";
+import { updateFaktumLagretVerdi, updateFaktumMedLagretVerdi, updateFaktumStateVerdi } from "./faktaUtils";
+import { Reducer } from "../reduxTypes";
+import { FaktumActionTypes } from "./faktaTypes";
 
 export interface FaktumState {
 	restStatus: string;
@@ -13,81 +10,11 @@ export interface FaktumState {
 	data: Faktum[];
 }
 
-export interface FaktumComponentProps {
-	fakta: Faktum[];
-}
-
 const initialState: FaktumState = {
 	restStatus: "",
 	progresjonPending: false,
 	data: []
 };
-
-export type FaktumActionTypes =
-	| LagreFaktum
-	| LagretFaktum
-	| OppdaterFaktumVerdi
-	| OpprettFaktum
-	| OpprettetFaktum
-	| SlettFaktum
-	| SlettetFaktum
-	| SetFaktaAction
-	| SetFaktumFailedAction
-	| SetFaktumIgnorert
-	| ResetFaktaAction;
-
-interface ResetFaktaAction {
-	type: FaktaActionTypeKeys.RESET_FAKTA;
-}
-
-interface OppdaterFaktumVerdi {
-	type: FaktumActionTypeKeys.OPPDATER_FAKTUM;
-	faktum: Faktum;
-}
-
-interface LagreFaktum {
-	type: FaktumActionTypeKeys.LAGRE_FAKTUM;
-	faktum: Faktum;
-}
-
-interface LagretFaktum {
-	type: FaktumActionTypeKeys.LAGRET_FAKTUM;
-	faktum: Faktum;
-}
-
-interface OpprettFaktum {
-	type: FaktumActionTypeKeys.OPPRETT_FAKTUM;
-}
-
-interface OpprettetFaktum {
-	type: FaktumActionTypeKeys.OPPRETTET_FAKTUM;
-	faktum: Faktum;
-}
-
-interface SlettFaktum {
-	type: FaktumActionTypeKeys.SLETT_FAKTUM;
-	faktumId: number;
-}
-
-interface SlettetFaktum {
-	type: FaktumActionTypeKeys.SLETTET_FAKTUM;
-}
-
-interface SetFaktaAction {
-	type: FaktaActionTypeKeys.SET_FAKTA;
-	fakta: Faktum[];
-}
-
-interface SetFaktumFailedAction {
-	type: FaktumActionTypeKeys.FEILET;
-	feilmelding: string;
-}
-
-interface SetFaktumIgnorert {
-	type: FaktumActionTypeKeys.IGNORER_FAKTUM;
-	faktum: Faktum;
-	ignorert: boolean;
-}
 
 const FaktumReducer: Reducer<FaktumState, FaktumActionTypes> = (
 	state = initialState,
